refactor: file structure refactoring
