+++ app/_features/home/create-room/index.ts
@@ -1,0 +1,1 @@
+export { default } from '@/_features/home/create-room/create-room';
