+++ app/_features/home/footer/index.ts
@@ -1,0 +1,1 @@
+export { default } from '@/_features/home/footer/footer';
