+++ app/_features/home/header/index.ts
@@ -1,0 +1,1 @@
+export { default } from '@/_features/home/header/header';
