+++ app/_features/home/join-room/index.ts
@@ -1,0 +1,1 @@
+export { default } from '@/_features/home/join-room/join-room';
