+++ app/_features/home/layout/index.ts
@@ -1,0 +1,1 @@
+export { default } from '@/_features/home/layout/layout';
