+++ app/_features/home/main/index.ts
@@ -1,0 +1,1 @@
+export { default } from '@/_features/home/main/main';
